Consolidate embedding tests
