+++ frontend/test/__support__/server-mocks/card.ts
@@ -74,7 +74,3 @@ export function setupCardPublicLinkEndpoints(cardId: CardId) {
     id: cardId,
   });
 }
-
-export function setupCardEmbeddable() {
-  fetchMock.get("path:/api/card/embeddable", []);
-}
