+++ frontend/test/__support__/server-mocks/dashboard.ts
@@ -51,7 +51,3 @@ export function setupDashboardPublicLinkEndpoints(dashboardId: DashboardId) {
     id: dashboardId,
   });
 }
-
-export function setupDashboardEmbeddable() {
-  fetchMock.get("path:/api/dashboard/embeddable", []);
-}
